kakaoAdapter / GlobalApplication
