+++ app/src/main/java/com/nimontoy/android/adapter/GlobalApplication.kt
@@ -7,48 +7,32 @@ import com.kakao.auth.*
 import org.koin.core.context.startKoin
 import org.koin.core.logger.EmptyLogger
 
-class GlobalApplication: Application() {
+object GlobalApplication: Application() {
+    @Volatile private var obj: GlobalApplication? = null
+    @Volatile private var currentActivity: Activity? = null
+
     override fun onCreate() {
         super.onCreate()
-        application = this
+        obj = this
         KakaoSDK.init(KakaoSDKAdapter())
 
         // start koin
         //startKoin(this, appModule, logger = EmptyLogger())
     }
 
-    companion object {
-        var application: GlobalApplication? = null
+
+   fun getGlobalApplicationContext(): GlobalApplication?{
+        return obj;
     }
 
-    internal inner class KakaoSDKAdapter : KakaoAdapter() {
-        override fun getSessionConfig(): ISessionConfig {
-            return object : ISessionConfig {
-                override fun getAuthTypes(): Array<AuthType> {
-                    return arrayOf(AuthType.KAKAO_TALK)
-                }
-
-                override fun isUsingWebviewTimer(): Boolean {
-                    return false
-                }
-
-                override fun isSecureMode(): Boolean {
-                    return false
-                }
-
-                override fun getApprovalType(): ApprovalType {
-                    return ApprovalType.INDIVIDUAL
-                }
-
-                override fun isSaveFormData(): Boolean {
-                    return true
-                }
-            }
-        }
-
-        override fun getApplicationConfig(): IApplicationConfig {
-            return IApplicationConfig { GlobalApplication.Companion.application }
-        }
+    fun getCurrentActivity(): Activity? {
+        return currentActivity;
     }
+
+    // Activity가 올라올때마다 Activity의 onCreate에서 호출해야 한다.
+    fun setCurrentActivity(currentActivity: Activity) {
+        this.currentActivity = currentActivity;
+    }
+
 }
